Day 3, Part 1, Revision 1
this feel extremely messy and gross but it produced a correct answer. i'd like to look at this in the morning and learn how to see if i can do something clever with binary math, and learn what swift has for that. there's probably a good bit of general code cleanup too that i didn't have the energy for right now.